requirements.txt
astroid==2.4.2
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pyaes==1.6.1
pyasn1==0.4.8
pylint==2.5.3
python-dotenv==0.14.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.18
Telethon==1.16.1
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.12.1
